update 12/6
added a close db function, modified the assignments table eliminating the grading_type attribute (we had already commented this out) and adjust the alligned tables to handle this. Modified fetchAssignmentsByID to handle the respected API fetch. Also, about to start testing the call to db for the array of objects.